Moved library out of node_modules to lib
Moved the node_modules/lib dir out to lib as this is going to be a
library package rather than the main application.

In addition, further work towards completing the primary use cases of
the package. I have a suspicion that I'll be making a further alteration
to the interfaces exposed via the main entry point of the library.
However, as this is 31st Dec, I wanted to commit something before the
start of 2015!

Happy new year!