Import key pair use case and started readme
I completed the work for the importKeyPair function. Importing a key
pair is the alternative to generating a key pair with a passphrase. This
will be useful for people wishing to bring their own keys.

In addition, I've made a start on the README. Bit by bit, I'll add in
the core documentation. Currently missing from the core use cases is the
use case to allow the application author to generate a user entity to be
used in other usecases.